Changed MQTT to use command in general, MQTT log levels settable via MQTT, coherent use of mutex utils, better mutex for ade7953, removed debug flags, discovered bug in arduinojson

+++ source/include/led.h
@@ -76,9 +76,9 @@ namespace Led {
 
     void resetToDefaults();
 
-    void setBrightness(uint32_t brightness);
-    uint32_t getBrightness();
-    inline bool isBrightnessValid(uint32_t brightness) { return brightness <= LED_MAX_BRIGHTNESS_PERCENT; }
+    void setBrightness(uint8_t brightness);
+    uint8_t getBrightness();
+    inline bool isBrightnessValid(uint8_t brightness) { return brightness <= LED_MAX_BRIGHTNESS_PERCENT; }
 
     // Pattern control functions
     void setPattern(LedPattern pattern, Color color, LedPriority priority = 1, uint64_t durationMs = 0);
